refactor code into a main.py to test the code and move everything else into src so it can be imported into other repos
